% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fs_cat_to_id}
\alias{fs_cat_to_id}
\title{Helper function that matches string categories to id's}
\usage{
fs_cat_to_id(category_id)
}
\arguments{
\item{category_id}{Must be a valid category string, regardless of case}
}
\value{
a vector of integers corresponding to valid figshare categories
}
\description{
Helper function that matches string categories to id's
}
\author{
Edmund Hart \email{edmund.m.hart@gmail.com}
}
\references{
\url{http://api.figshare.com}
}

