% Generated by roxygen2 (4.0.1): do not edit by hand
\name{fs_create_author}
\alias{fs_create_author}
\title{Creates a figshare author}
\usage{
fs_create_author(full_name, session = fs_get_auth(), debug = FALSE)
}
\arguments{
\item{full_name}{full name of the author to create}

\item{session}{(optional) the authentication credentials from \code{\link{fs_auth}}. If not provided, will attempt to load from cache as long as figshare_auth has been run.}

\item{debug}{return PUT request visibly?}
}
\value{
author ID numbers
}
\description{
Creates a figshare author
}
\examples{
\dontrun{
fs_create_author("Benjamin Franklin")
}
}
\author{
Carl Boettiger \email{cboettig@gmail.com}
}
\references{
\url{http://api.figshare.com}
}
\seealso{
\code{\link{fs_auth}}
}

