% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/fs_details.R
\name{fs_details}
\alias{fs_details}
\title{Get details for an article}
\usage{
fs_details(article_id, mine = is_mine(article_id), session = fs_get_auth(),
  show_versions = FALSE, version = NULL, debug = FALSE)
}
\arguments{
\item{article_id}{number}

\item{mine}{logical (default FALSE). Set to true to see article details for your own non-public articles}

\item{session}{the authentication credentials from \code{\link{fs_auth}}}

\item{show_versions}{logical, show what versions are available}

\item{version}{show a given version number}

\item{debug}{logical, enable debugging mode?}
}
\description{
Get details for an article
}
\examples{
\dontrun{
fs_details(138)
}
}
\author{
Carl Boettiger \email{cboettig@gmail.com}
}
\references{
\url{http://api.figshare.com}
}
\seealso{
\code{\link{fs_auth}}
}

