% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_embed.R
\name{fs_image_url}
\alias{fs_image_url}
\title{get image url}
\usage{
fs_image_url(id, debug = FALSE)
}
\arguments{
\item{id}{a (public) figshare figure id number}

\item{debug}{logical, enable debugging mode?}
}
\value{
a url to the image file
}
\description{
get image url
}
\details{
this is currently an unstable hack of html parsing
}
