% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rfigshare-package.R
\docType{package}
\name{rfigshare-package}
\alias{rfigshare-package}
\alias{rfigshare}
\title{An R Interface to Figshare}
\description{
An R Interface to Figshare
}
\section{About}{

word and stuff ...
}

\keyword{package}
