% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fs_auth.R
\name{fs_auth}
\alias{fs_auth}
\title{Figshare authentication}
\usage{
fs_auth(token, force = FALSE)
}
\arguments{
\item{token}{(character) A Figshare personal access token. Required.
See Details.}
}
\value{
sets your token as an env variable for the current R session,
returns nothing
}
\description{
Figshare authentication
}
\details{
Instructions:
\itemize{
 \item Log in
 \item Go to "Applications" from the drop down menu in top right
 \item Scroll to bottom, find "Personal Tokens" section
 \item Click on "Create Personal Token"
}
}
\examples{
\dontrun{
fs_auth("aasdfasfasdfasdfasfasfds")
}
}
\author{
Carl Boettiger \email{cboettig@gmail.com}
}
\references{
\url{http://api.figshare.com}
}

