% Generated by roxygen2 (4.1.1.9000): do not edit by hand
% Please edit documentation in R/fs_category_list.R
\name{fs_category_list}
\alias{fs_category_list}
\title{List all categories}
\usage{
fs_category_list(debug = FALSE)
}
\arguments{
\item{debug}{enable debugging}
}
\value{
a table of all the categories
}
\description{
List all categories
}
\examples{
\dontrun{
fs_categories_list()
}
}
\references{
\url{http://api.figshare.com}
}

