\name{fs_ids}
\alias{fs_ids}
\title{Get a list of article id numbers from a search return}
\usage{
  fs_ids(object)
}
\arguments{
  \item{object}{the output of a search}
}
\value{
  a list of article id numbers
}
\description{
  Get a list of article id numbers from a search return
}
\examples{
\dontrun{
figshare_category()
}
}
\author{
  Carl Boettiger \email{cboettig@gmail.com}
}
\references{
  \url{http://api.figshare.com}
}

